requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0
django-allauth==0.61.1
django-ckeditor-5==0.2.10
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-google-fonts==0.0.3
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-solo==2.1.0
django-urlconfchecks==0.10.0
idna==3.6
mypy-extensions==1.0.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.6.0
shellingham==1.5.4
sqlparse==0.4.4
tinycss2==1.2.1
typer==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
